Move sqlite internals to separate sql crate
